requirements.txt
alembic==0.9.6
beautifulsoup4==4.6.0
bs4==0.0.1
click==6.7
fake-useragent==0.1.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
gunicorn
